move test into right folder
